% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACC_lncRNA_Exp}
\alias{ACC_lncRNA_Exp}
\title{lncRNA expression data}
\format{
ACC_lncRNA_Exp: A matrix object with 79 ACC 
samples (rows) and 4366 lncRNAs (columns).
}
\description{
lncRNA expression data
}
\details{
The matched ncRNA (including miRNA, lncRNA and pseudogene)
and mRNA expression data in Adrenocortical carcinoma (ACC) are obtained 
from The Cancer Genome Atlas (TCGA, http://cancergenome.nih.gov/).
}
