% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACC_pseudogene_Exp}
\alias{ACC_pseudogene_Exp}
\title{pseudogene expression data}
\format{
ACC_pseudogene_Exp: A matrix object with 79 ACC 
samples (rows) and 11952 pseudogenes (columns).
}
\description{
pseudogene expression data
}
\details{
The matched ncRNA (including miRNA, lncRNA and pseudogene)
and mRNA expression data in Adrenocortical carcinoma (ACC) are obtained 
from The Cancer Genome Atlas (TCGA, http://cancergenome.nih.gov/).
}
