% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRTarget_lncR_vs_mR}
\alias{miRTarget_lncR_vs_mR}
\title{miRNA-lncRNA and miRNA-mRNA interactions}
\format{
miRTarget_lncR_vs_mR: A data.frame object with
975621 miRNA-lncRNA and miRNA-mRNA interactions.
}
\description{
miRNA-lncRNA and miRNA-mRNA interactions
}
\details{
The putative miRNA-mRNA interactions (762,540) are acquired 
from miRTarBase v9.0 and TarBase v8.0, and the putative miRNA-lncRNA 
interactions (225,063) are obtained from LncBase v2.0 and NPInter 4.0.
}
\references{
Huang H, Lin Y-C, Cui S, Huang Y, Tang Y, Xu J, et al. 
miRTarBase update 2022: an informative resource for experimentally 
validated miRNA-target interactions. Nucleic Acids Res. 
2022;50: D222–D230. doi:10.1093/nar/gkab1079

Karagkouni D, Paraskevopoulou MD, Chatzopoulos S, Vlachos IS, 
Tastsoglou S, Kanellos I, et al. DIANA-TarBase v8: a decade-long 
collection of experimentally supported miRNA-gene interactions. 
Nucleic Acids Res. 2018;46: D239–D245. doi:10.1093/nar/gkx1141

Paraskevopoulou MD, Vlachos IS, Karagkouni D, Georgakilas G, 
Kanellos I, Vergoulis T, et al. DIANA-LncBase v2: indexing microRNA 
targets on non-coding transcripts. Nucleic Acids Res. 2016;44: D231-238. 
doi:10.1093/nar/gkv1270

Teng X, Chen X, Xue H, Tang Y, Zhang P, Kang Q, et al. 
NPInter v4.0: an integrated database of ncRNA interactions. Nucleic Acids Res. 
2020;48: D160–D165. doi:10.1093/nar/gkz969
}
