% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRTarget_pseudo_vs_mR}
\alias{miRTarget_pseudo_vs_mR}
\title{miRNA-pseudogene and miRNA-mRNA interactions}
\format{
miRTarget_pseudo_vs_mR: A data.frame object with
854141 miRNA-pseudogene and miRNA-mRNA interactions.
}
\description{
miRNA-pseudogene and miRNA-mRNA interactions
}
\details{
The putative miRNA-mRNA interactions (762,540) are acquired 
from miRTarBase v9.0 and TarBase v8.0, and the putative miRNA-pseudogene 
interactions (91,619) are from ENCORI (the previous version is starBase).
}
\references{
Huang H, Lin Y-C, Cui S, Huang Y, Tang Y, Xu J, et al. 
miRTarBase update 2022: an informative resource for experimentally 
validated miRNA-target interactions. Nucleic Acids Res. 
2022;50: D222–D230. doi:10.1093/nar/gkab1079

Karagkouni D, Paraskevopoulou MD, Chatzopoulos S, Vlachos IS, 
Tastsoglou S, Kanellos I, et al. DIANA-TarBase v8: a decade-long 
collection of experimentally supported miRNA-gene interactions. 
Nucleic Acids Res. 2018;46: D239–D245. doi:10.1093/nar/gkx1141

Li JH, Liu S, Zhou H, Qu LH, Yang JH. 
starBase v2.0: decoding miRNA-ceRNA, miRNA-ncRNA and protein-RNA 
interaction networks from large-scale CLIP-Seq data. Nucleic Acids Res. 
2014;42: D92-97. doi:10.1093/nar/gkt1248
}
