% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{module.classify}
\alias{module.classify}
\title{module.classify}
\usage{
module.classify(
  Exp,
  tumor_type,
  Modulelist,
  method = "br",
  base.algorith = "SVM",
  cv.folds = 10,
  cv.sampling = "stratified",
  cv.seed = 12345
)
}
\arguments{
\item{Exp}{Gene expression data, rows are samples, columns are genes.}

\item{tumor_type}{Molecular subtype information of a tumor.}

\item{Modulelist}{A list of gene modules.}

\item{method}{The multi-label classification method. It also accepts the name of the method as a string.}

\item{base.algorith}{The base algorithm of the multi-label classification method.}

\item{cv.folds}{Number of folds (Default: 10).}

\item{cv.sampling}{The method to split the data. (Default: "stratified")}

\item{cv.seed}{An optional integer used to set the seed.}
}
\value{
Matrix object: Multi-label classification results of a tumor.
}
\description{
Evaluating the performance of gene list in classifying tumor types.
}
\references{
Tsoumakas G, Katakis I, Vlahavas I. Mining multi-label data. 
Data mining and knowledge discovery handbook Springer, Boston, MA. Springer, Boston, MA; 2009. pp. 667–685.

Rivolli A, Carvalho ACPLF de. The utiml package: multi-label 
classification in R. The R Journal. 2018;10: 24–37. doi:10.32614/RJ-2018-041

Chang CC, Lin CJ. LIBSVM: a library for support vector machines. 
ACM Transactions on Intelligent Systems and Technology. 2011;2: 1–27.

Meyer D, Dimitriadou E, Hornik K, Weingessel A, Leisch F, Chang C-C, et al. 
e1071: misc functions of the department of statistics, probability theory group (Formerly: E1071), 
TU Wien. Available: https://CRAN.R-project.org/package=e1071
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2}
}
