% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{Random_net_parallel}
\alias{Random_net_parallel}
\title{Random_net_parallel}
\usage{
Random_net_parallel(
  obser_path,
  obser_density,
  nodes.num,
  edges.num,
  perm = 100,
  directed = FALSE,
  num.cores = 32
)
}
\arguments{
\item{obser_path}{Observed characteristics path length of a biological network.}

\item{obser_density}{Observed density of a biological network.}

\item{nodes.num}{The number of nodes in a biological network.}

\item{edges.num}{The number of edges in a biological network.}

\item{perm}{The number of permutations for generating random networks.}

\item{directed}{A logical value, FALSE for undirected network and TRUE for directed network.}

\item{num.cores}{The number of cores used for parallel calculation.}
}
\value{
A vector: The significance -log10(p-value) of characteristics path length and density of a biological network.
}
\description{
Calculating the significance p-value of topological characteristics in biological networks.
}
\examples{
# NOT RUN
# library(SPONGE)
# data(ACC)
# num.cores <- 2
# cores <- makeCluster(num.cores)
# registerDoParallel(cores)
# pre_null_model_ACC <- sponge_build_null_model(number_of_datasets = 1e+03, 
# number_of_samples = nrow(ACC_miRNA_Exp))
# ceRNet_lncR_vs_mR_ACC <- SC(ACC_miRNA_Exp, ACC_lncRNA_Exp, ACC_mRNA_Exp, miRTarget_lncR_vs_mR, null_model = pre_null_model_ACC)
# ceRNet_lncR_vs_mR_ACC_graph <- make_graph(c(t(ceRNet_lncR_vs_mR_ACC[, 1:2])), directed = FALSE)
# ceRNet_lncR_vs_mR_ACC_path <- mean_distance(ceRNet_lncR_vs_mR_ACC_graph)
# ceRNet_lncR_vs_mR_ACC_density <- edge_density(ceRNet_lncR_vs_mR_ACC_graph)
# ceRNet_lncR_vs_mR_ACC_random <- Random_net_parallel(ceRNet_lncR_vs_mR_ACC_path, ceRNet_lncR_vs_mR_ACC_density, length(V(ceRNet_lncR_vs_mR_ACC_graph)), length(E(ceRNet_lncR_vs_mR_ACC_graph)))
# ceRNet_lncR_vs_mR_ACC_path_p <- ceRNet_lncR_vs_mR_ACC_random[1]
# ceRNet_lncR_vs_mR_ACC_density_p <- ceRNet_lncR_vs_mR_ACC_random[2]

}
\references{
Csardi G, Nepusz T. The igraph software package for complex network research. 
InterJournal, Complex Systems. 2006, 1695.
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
