% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{SC}
\alias{SC}
\title{SC}
\usage{
SC(
  miRExp,
  ceRExp,
  mRExp,
  miRTarget,
  minSharedmiR = 1,
  padjustvaluecutoff = 0.05,
  padjustmethod = "none",
  poscorcutoff = 0,
  senscorcutoff = 0,
  null_model
)
}
\arguments{
\item{miRExp}{miRNA expression data, rows are samples, columns are miRNAs.}

\item{ceRExp}{ceRNA (lncRNAs or pseudogenes) expression data, rows are samples, columns are ceRNAs.}

\item{mRExp}{mRNA expression data, rows are samples, columns are mRNAs.}

\item{miRTarget}{Putative miRNA-target interactions.}

\item{minSharedmiR}{The minimum number of shared miRNAs between ceRNAs and mRNAs.}

\item{padjustvaluecutoff}{A cutoff value of adjusted p-values.}

\item{padjustmethod}{Adjusted method of p-values, can select one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{poscorcutoff}{A cutoff value of positive correlation.}

\item{senscorcutoff}{A cutoff value of sensitivity correlation.}

\item{null_model}{Pre-computed null model. Users can build the null model using "sponge_build_null_model" function in SPONGE R package.}
}
\value{
Matrix object: Predicted ceRNA interactions.
}
\description{
Identifying ceRNA interactions using Sensitivity Correlation (SC) method and null model.
}
\examples{
# NOT RUN
# library(SPONGE)
# data(ACC)
# num.cores <- 2
# cores <- makeCluster(num.cores)
# registerDoParallel(cores)
# pre_null_model_ACC <- sponge_build_null_model(number_of_datasets = 1e+03, 
# number_of_samples = nrow(ACC_miRNA_Exp))
# ceRNet_lncR_vs_mR_ACC <- SC(ACC_miRNA_Exp, ACC_lncRNA_Exp, ACC_mRNA_Exp, miRTarget_lncR_vs_mR, null_model = pre_null_model_ACC)

}
\references{
Langfelder P, Horvath S. WGCNA: an R package for weighted 
correlation network analysis. BMC Bioinformatics. 2008, 9:559.

List M, Dehghani Amirabad A, Kostka D, Schulz MH. 
Large-scale inference of competing endogenous RNA networks with 
sparse partial correlation. Bioinformatics. 2019;35: i596–i604.

Paci P, Colombo T, Farina L. Computational analysis identifies 
a sponge interaction network between long non-coding RNAs and messenger RNAs 
in human breast cancer. BMC Syst Biol. 2014;8: 83. doi:10.1186/1752-0509-8-83
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
