% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{SCOM}
\alias{SCOM}
\title{SCOM}
\usage{
SCOM(
  ceRnet,
  ncRExpData,
  mRExpData,
  minSharedmR = 1,
  padjustvaluecutoff = 0.05,
  padjustmethod = "none",
  poscorcutoff = 0,
  senscorcutoff = 0,
  null_model
)
}
\arguments{
\item{ceRnet}{Putative ceRNA interactions.}

\item{ncRExpData}{ncRNA expression data, rows are samples, columns are ncRNAs.}

\item{mRExpData}{mRNA expression data, rows are samples, columns are mRNAs.}

\item{minSharedmR}{The minimum number of shared competed mRNAs between ceRNAs.}

\item{padjustvaluecutoff}{A cutoff value of adjusted p-values.}

\item{padjustmethod}{Adjusted method of p-values, can select one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{poscorcutoff}{A cutoff value of positive correlation.}

\item{senscorcutoff}{A cutoff value of sensitivity correlation.}

\item{null_model}{Pre-computed null model. Users can build the null model using "sponge_build_null_model" function in SPONGE R package.}
}
\value{
Matrix object: Predicted ncRNA synergistic competition relationships.
}
\description{
Identifying ncRNA synergistic competition using Synergistic COMpetition (SCOM) method and null model.
}
\examples{
# NOT RUN
# library(SPONGE)
# data(ACC)
# num.cores <- 2
# cores <- makeCluster(num.cores)
# registerDoParallel(cores)
# pre_null_model_ACC <- sponge_build_null_model(number_of_datasets = 1e+03, 
# number_of_samples = nrow(ACC_miRNA_Exp))
# ceRNet_lncR_vs_mR_ACC <- SC(ACC_miRNA_Exp, ACC_lncRNA_Exp, ACC_mRNA_Exp, miRTarget_lncR_vs_mR, null_model = pre_null_model_ACC)
# ceRNet_pseudo_vs_mR_ACC <- SC(ACC_miRNA_Exp, ACC_pseudogene_Exp, ACC_mRNA_Exp, miRTarget_pseudo_vs_mR, null_model = pre_null_model_ACC)
# ceRNet_ACC <- rbind(ceRNet_lncR_vs_mR_ACC, ceRNet_pseudo_vs_mR_ACC)
# SCOMNet_ACC <- SCOM(ceRNet_ACC, cbind(ACC_lncRNA_Exp, ACC_pseudogene_Exp), ACC_mRNA_Exp, null_model = pre_null_model_ACC)

}
\references{
Langfelder P, Horvath S. WGCNA: an R package for weighted 
correlation network analysis. BMC Bioinformatics. 2008, 9:559.

List M, Dehghani Amirabad A, Kostka D, Schulz MH. 
Large-scale inference of competing endogenous RNA networks with 
sparse partial correlation. Bioinformatics. 2019;35: i596–i604.

Paci P, Colombo T, Farina L. Computational analysis identifies 
a sponge interaction network between long non-coding RNAs and messenger RNAs 
in human breast cancer. BMC Syst Biol. 2014;8: 83. doi:10.1186/1752-0509-8-83
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
