% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{Sim.ceRNet}
\alias{Sim.ceRNet}
\title{Sim.ceRNet}
\usage{
Sim.ceRNet(net1, net2, directed = FALSE)
}
\arguments{
\item{net1}{List object, the first list of networks.}

\item{net2}{List object, the second list of networks.}

\item{directed}{Logical value, network directed (TRUE) or undirected (FALSE).}
}
\value{
Matrix object: A similarity matrix between two list of networks.
}
\description{
Calculating similarity matrix between two list of networks
}
\examples{
library(igraph) 
net1 <- list(as_data_frame(sample_k_regular(10, 2)), as_data_frame(sample_k_regular(10, 2))) 
net2 <- list(as_data_frame(sample_k_regular(10, 3)), as_data_frame(sample_k_regular(10, 3))) 
Sim.net <- Sim.ceRNet(net1, net2)

}
\references{
Tucker CM, Cadotte MW, Carvalho SB, et al. A guide to phylogenetic metrics for conservation, community ecology and macroecology. Biol Rev Camb Philos Soc. 2017;92(2):698-715.
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
