% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOM.R
\name{Sim.hub}
\alias{Sim.hub}
\title{Sim.hub}
\usage{
Sim.hub(hub1, hub2)
}
\arguments{
\item{hub1}{List object, the first list of hub.}

\item{hub2}{List object, the second list of hub.}
}
\value{
Matrix object: A similarity matrix between two list of hubs.
}
\description{
Calculating similarity matrix between two list of hubs.
}
\examples{
library(igraph) 
hub1 <- list(c("ncRNA1", "ncRNA2", "ncRNA3", "ncRNA4"), c("ncRNA1", "ncRNA2", "ncRNA4", "ncRNA5")) 
hub2 <- list(c("ncRNA1", "ncRNA2", "ncRNA5", "ncRNA6"), c("ncRNA1", "ncRNA3", "ncRNA4", "ncRNA6"))
Sim_hub <- Sim.hub(hub1, hub2) 

}
\references{
Tucker CM, Cadotte MW, Carvalho SB, et al. A guide to phylogenetic metrics for conservation, community ecology and macroecology. Biol Rev Camb Philos Soc. 2017;92(2):698-715.
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
